Move clio-get to packageConfig
