+++ cli/commands/deps_commands/add.js
@@ -1,4 +1,4 @@
-const { get } = require("../../../internals/get/clio-get");
+const { get } = require("../../../package/packageConfig");
 
 exports.command = "add <url>";
 exports.desc = "Add a new dependency";
